[refactor] StoryDetail: move to keep ViewModel at NavGraph; create UIEvent to pass lambdas
